typescript(vx-demo): rewrite Geo demos in TypeScript
