+++ packages/vx-demo/src/components/gallery.js
@@ -27,8 +27,8 @@ import Cluster from './tiles/dendrogram';
 import Voronoi from './tiles/voronoi';
 import Legends from './tiles/legends';
 import StatsPlot from './tiles/statsplot';
-import GeoMercator from './tiles/geo-mercator';
-import GeoCustom from './tiles/geo-custom';
+import GeoMercator from './tiles/Geo-Mercator.tsx';
+import GeoCustom from './tiles/Geo-Custom.tsx';
 import Network from './tiles/Network.tsx';
 import Streamgraph from './tiles/streamgraph';
 import Pack from './tiles/Pack.tsx';
