+++ packages/vx-demo/src/components/tiles/Geo-Mercator.tsx
@@ -6,6 +6,7 @@ import topology from '../../static/vx-geo/world-topo.json';
 
 const bg = '#f9f7e8';
 
+// @ts-ignore
 const world = topojson.feature(topology, topology.objects.units);
 const color = scaleQuantize({
   domain: [
@@ -25,27 +26,33 @@ export default ({ width, height, events = false }) => {
   return (
     <svg width={width} height={height}>
       <rect x={0} y={0} width={width} height={height} fill={bg} rx={14} />
-      <Mercator data={world.features} scale={scale} translate={[centerX, centerY + 50]}>
+      <Mercator<{
+        type: 'Feature';
+        geometry: { coordinates: [number, number][][][]; type: 'MultiPolygon' };
+        properties: { name: string };
+        id: string;
+      }>
+        data={world.features}
+        scale={scale}
+        translate={[centerX, centerY + 50]}
+      >
         {mercator => {
           return (
             <g>
               <Graticule graticule={g => mercator.path(g)} stroke="rgba(33,33,33,0.05)" />
-              {mercator.features.map((feature, i) => {
-                const { feature: f } = feature;
-                return (
-                  <path
-                    key={`map-feature-${i}`}
-                    d={feature.path}
-                    fill={color(f.geometry.coordinates.length)}
-                    stroke={bg}
-                    strokeWidth={0.5}
-                    onClick={() => {
-                      if (!events) return;
-                      alert(`Clicked: ${f.properties.name} (${f.id})`);
-                    }}
-                  />
-                );
-              })}
+              {mercator.features.map(({ feature, path }, i) => (
+                <path
+                  key={`map-feature-${i}`}
+                  d={path}
+                  fill={color(feature.geometry.coordinates.length)}
+                  stroke={bg}
+                  strokeWidth={0.5}
+                  onClick={() => {
+                    if (!events) return;
+                    alert(`Clicked: ${feature.properties.name} (${feature.id})`);
+                  }}
+                />
+              ))}
             </g>
           );
         }}
